Add login functionality
In short, login works by giving out JWTs (JSON web tokens) from the
endpoint `/b/login`. Access tokens can then be stored and used for
future requests.

Passwords are hased in the database with Argon2. This is an acceptable
hash for use in passwords because it's relatively slow.

As far as the frontend goes, I added a global token context to store the
access token. In the future it might be wise to have other global state
like information on the current user.

OAuth should Just Work™ but I have not tested it yet. I expect it to
need a few tweaks to be functional.